backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
openai>=1.8.0
pillow>=10.0.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
uvicorn>=0.25.0
starlette>=0.27.0
tiktoken>=0.5.0